refactor: move 'createOrJoin' to separate component to allow stateful rendering of welcome screen
