+++ src/components/index.js
@@ -1,3 +1,4 @@
+import JoinList from './JoinList';
 import Modal from './Modal';
 import ShareToken from './ShareToken';
 import GlobalStyles, { primary, accent, neutral } from './General/globalStyles';
@@ -6,6 +7,7 @@ import { theme } from './General/globalStyles';
 
 // Then you can export all of the components from here
 export {
+  JoinList,
   Modal,
   ShareToken,
   GlobalStyles,
